Avoid insufficient casting (#989)

+++ contracts/truefi2/TrueLender2.sol
@@ -384,11 +384,11 @@ contract TrueLender2 is ITrueLender2, UpgradeableClaimable {
 
         // loop through loans and sum amount borrowed accounting for precision
         ITrueFiPool2[] memory pools = factory.getSupportedPools();
-        for (uint8 i = 0; i < pools.length; i++) {
+        for (uint256 i = 0; i < pools.length; i++) {
             ITrueFiPool2 pool = pools[i];
             uint256 poolPrecision = uint256(10)**ITrueFiPool2WithDecimals(address(pool)).decimals();
             ILoanToken2[] memory _loans = poolLoans[pool];
-            for (uint8 j = 0; j < _loans.length; j++) {
+            for (uint256 j = 0; j < _loans.length; j++) {
                 ILoanToken2 loan = _loans[j];
                 if (address(loan.borrower()) == borrower) {
                     uint256 loanValue = loan.value(loan.balanceOf(address(this)));
